server/requirements.txt
fastapi
uvicorn
python-dotenv
pandas
openpyxl
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pydantic
requests
